+++ apps/rest-api-server/src/network.ts
@@ -1,4 +1,0 @@
-import { env } from "@blobscan/env";
-import { getNetwork } from "@blobscan/network-blob-config";
-
-export const network = getNetwork(env.CHAIN_ID);