+++ packages/api/src/routers/blob/getAll.ts
@@ -47,9 +47,7 @@ export const getAll = publicProcedure
   .use(withExpands)
   .output(outputSchema)
   .query(
-    async ({
-      ctx: { filters, expands, pagination, prisma, count, network },
-    }) => {
+    async ({ ctx: { filters, expands, pagination, prisma, count, chain } }) => {
       const {
         blockFilters = {},
         blockType,
@@ -89,9 +87,7 @@ export const getAll = publicProcedure
         ],
         ...pagination,
       });
-      const countOp = count
-        ? countBlobs(prisma, network.id, filters)
-        : undefined;
+      const countOp = count ? countBlobs(prisma, chain.id, filters) : undefined;
 
       const [prismaBlobsOnTxs, totalBlobs] = await Promise.all([
         blobsOnTransactinonsOp,