+++ packages/db/prisma/seed/network.ts
@@ -1,3 +1,0 @@
-import { getNetwork } from "@blobscan/network-blob-config";
-
-export const network = getNetwork("mainnet");